- Moved more scanning related classes into the scanning package
