+++ src/burp/BurpExtender.java
@@ -6,9 +6,13 @@ import burp.model.param.OfferParamGuess;
 import burp.model.param.ParamGrabber;
 import burp.model.scanning.BulkScan;
 import burp.model.scanning.BulkScanLauncher;
+import burp.model.scanning.FatGetScan;
 import burp.model.scanning.GrabScan;
 import burp.model.scanning.NormalisedParamScan;
 import burp.model.scanning.NormalisedPathScan;
+import burp.model.scanning.PortDosScan;
+import burp.model.scanning.RailsUtmScan;
+import burp.model.scanning.UnkeyedParamScan;
 import burp.model.utilities.Utilities;
 import burp.view.ConfigMenu;
 import burp.api.montoya.BurpExtension;
@@ -144,10 +148,10 @@ public static SettingsBox guessSettings;
         BulkScanLauncher launcher = new BulkScanLauncher(BulkScan.scans, utilities);
         
         new HeaderPoison("Header poison", utilities, launcher);
-        new PortDOS("port-DoS", utilities, launcher);
+        new PortDosScan("port-DoS", utilities, launcher);
         //new ValueScan("param-value probe");
         new UnkeyedParamScan("Unkeyed param", utilities, launcher);
-        new FatGet("fat GET", utilities, launcher);
+        new FatGetScan("fat GET", utilities, launcher);
         new NormalisedParamScan("normalised param", utilities, launcher);
         new NormalisedPathScan("normalised path", utilities, launcher);
         new RailsUtmScan("rails param cloaking scan", utilities, launcher);
